adding orm

+++ src/cor/cor.hpp
@@ -7,12 +7,12 @@
 
 using i8_t = char;
 using i16_t = short;
-using i32_t = long;
+using i32_t = int;
 using i64_t = long long;
 
 using ui8_t = unsigned char;
 using ui16_t = unsigned short;
-using ui32_t = unsigned long;
+using ui32_t = unsigned int;
 using ui64_t = unsigned long long;
 
 using uimax_t = ui32_t;
